Create blockchain query services entities
