+++ blockchain_query_services/common/transaction_published_service.py
@@ -1,4 +1,4 @@
-from blockchain_query_services.blockchain_query_services_dependency_injection import (
+from blockchain_query_services.services.blockchain_query_services_dependency_injection import (
     transaction_info_service,
 )
