+++ blockchain_query_services/entities/transaction_info_service/transaction_output_bo.py
@@ -1,0 +1,13 @@
+from pydantic import BaseModel
+
+
+class TransactionOutputBO(BaseModel):
+    address: str
+    index: int
+    value: int
+
+    @staticmethod
+    def from_vout(vout: dict, index: int) -> "TransactionOutputBO":
+        return TransactionOutputBO(
+            address=vout["scriptpubkey_address"], index=index, value=vout["value"]
+        )
