+++ transactions/publication_services/trigger_wrong_hash_challenge_transaction_service.py
@@ -1,21 +1,3 @@
-from bitvmx_protocol_library.config import common_protocol_properties
-from bitvmx_protocol_library.enums import BitcoinNetwork
-
-if common_protocol_properties.network == BitcoinNetwork.MUTINYNET:
-    from blockchain_query_services.mutinynet_api.services.broadcast_transaction_service import (
-        BroadcastTransactionService,
-    )
-    from blockchain_query_services.mutinynet_api.services.transaction_info_service import (
-        TransactionInfoService,
-    )
-elif common_protocol_properties.network == BitcoinNetwork.TESTNET:
-    from blockchain_query_services.testnet_api.services import (
-        TransactionInfoService,
-    )
-elif common_protocol_properties.network == BitcoinNetwork.MAINNET:
-    from blockchain_query_services.mainnet_api.services.transaction_info_service import (
-        TransactionInfoService,
-    )
 from scripts.services.trigger_generic_challenge_script_generator_service import (
     TriggerGenericChallengeScriptGeneratorService,
 )
@@ -26,8 +8,6 @@ from winternitz_keys_handling.services.generate_witness_from_input_nibbles_servi
 
 class TriggerWrongHashChallengeTransactionService:
     def __init__(self, verifier_private_key):
-        self.transaction_info_service = TransactionInfoService()
-        self.broadcast_transaction_service = BroadcastTransactionService()
         self.verifier_challenge_execution_script_generator_service = (
             TriggerGenericChallengeScriptGeneratorService()
         )
